+++ back/src/model/repositories/users_repository.py
@@ -3,6 +3,8 @@ from src.model.entities.users import User
 from .interfaces.iusers_repository import IUsersRepository
 
 class UsersRepository(IUsersRepository):
+
+    
     def create_user(self, nome: str, email: str, senha: str, is_admin: bool = False) -> None:
         with DBConnectionHandler() as db:
             try:
@@ -13,6 +15,7 @@ class UsersRepository(IUsersRepository):
                 db.session.rollback()
                 raise exception
                 
+
     def get_user_by_id(self, user_id: int) -> User:
         with DBConnectionHandler() as db:
             user = (
@@ -23,6 +26,7 @@ class UsersRepository(IUsersRepository):
             )
             return user
         
+
     def list_users(self) -> list[User]:
         with DBConnectionHandler() as db:
             users = (
@@ -32,6 +36,7 @@ class UsersRepository(IUsersRepository):
             )
             return users
         
+
     def get_user_by_email(self, user_email: str) -> User:
         with DBConnectionHandler() as db:
             user = (
@@ -42,33 +47,30 @@ class UsersRepository(IUsersRepository):
             )
             return user
     
+
     def update_user(self, user_id: int, nome: str, email: str, senha: str, is_admin: bool) -> None:
         with DBConnectionHandler() as db:
             try:
-                user = (
-                    db.session
-                    .query(User)
-                    .filter(User.id == user_id)
-                    .one_or_none()
-                )
+                user = self.get_user_by_id(user_id)     
+                if not user:
+                    raise Exception("User not found!")      
                 user.nome = nome
                 user.email = email
                 user.senha = senha
                 user.is_admin = is_admin
+                db.session.add(user)
                 db.session.commit()
             except Exception as exception:
                 db.session.rollback()
                 raise exception
     
+
     def delete_user(self, user_id: int) -> None:
         with DBConnectionHandler() as db:
             try:
-                user = (
-                    db.session
-                    .query(User)
-                    .filter(User.id == user_id)
-                    .one_or_none()
-                )
+                user = self.get_user_by_id(user_id)
+                if not user:
+                    raise Exception("User not found!")      
                 db.session.delete(user)
                 db.session.commit()
             except Exception as exception:
